requirements.txt
pytest~=8.3.3
selenium==4.26.1
allure-pytest~=2.13.5
